refactor: split netinit.py into files
