+++ python/esxi-netinit/doc/example.json
