+++ python/esxi-netinit/netinit/nic.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class NIC:
+    name: str
+    status: str
+    link: str
+    mac: str
