+++ python/esxi-netinit/netinit/route.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Route:
+    gateway: str
+    netmask: str
+    network: str
+
+    def is_default(self):
+        return self.network == "0.0.0.0" and self.netmask == "0.0.0.0"
