+++ python/esxi-netinit/netinit/service.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class Service:
+    address: str
+    type: str
