+++ python/esxi-netinit/pyproject.toml
@@ -38,11 +38,16 @@ target-version = "py310"
     "S104",     # false positive on binding to all ifaces
 ]
 "tests/test_esxconfig.py" = [
-  "E501" # esxcli outputs with long lines
 ]
 "tests/test_esxhost.py" = [
   "E501" # esxcli outputs with long lines
 ]
-"netinit.py" = [
+"tests/test_niclist.py" = [
+  "E501" # esxcli outputs with long lines
+]
+"netinit/route.py" = [
   "S104",     # false positive on binding to all ifaces
 ]
+
+[tool.poetry.scripts]
+netinit = "netinit.main:main"
